Fix property ordering to go all / globals / groups / hosts. (lowest to highest priority)
Jira-Issue: OPENSTACK-295

+++ kollacli/common/inventory.py
@@ -576,7 +576,7 @@ class Inventory(object):
             subservice.remove_groupname(groupname)
 
         group_vars = os.path.join(get_group_vars_dir(), groupname)
-        if os.path.exists(group_vars):
+        if os.path.exists(group_vars) and groupname != '__RESERVED__':
             os.remove(group_vars)
 
         if groupname in self._groups:
